resolution de conflit et mise a jours de la branche principal

+++ src/main/java/oumaima_nezha_mehdy/zelda/Univers/Mur.java
@@ -2,7 +2,7 @@ package oumaima_nezha_mehdy.zelda.Univers;
 
 
 public class Mur extends Bloc {
-    public Mur(int x,int y) {
+    public Mur(double x,double y) {
         super(false,x,y);
 
     }
